Move InputStreamGenerator to jqf-fuzz
