+++ examples/pom.xml
@@ -67,7 +67,6 @@
         <dependency>
             <groupId>edu.berkeley.cs.jqf</groupId>
             <artifactId>jqf-fuzz</artifactId>
-            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.apache.struts</groupId>
