+++ examples/src/main/java/edu/berkeley/cs/jqf/examples/common/ArbitraryLengthStringGenerator.java
@@ -35,6 +35,7 @@ import com.pholser.junit.quickcheck.generator.GenerationStatus;
 import com.pholser.junit.quickcheck.generator.Generator;
 import com.pholser.junit.quickcheck.generator.Size;
 import com.pholser.junit.quickcheck.random.SourceOfRandomness;
+import edu.berkeley.cs.jqf.fuzz.junit.quickcheck.InputStreamGenerator;
 
 /**
  * @author Rohan Padhye
