+++ examples/src/main/resources/META-INF/services/com.pholser.junit.quickcheck.generator.Generator
@@ -1,4 +1,3 @@
-edu.berkeley.cs.jqf.examples.common.InputStreamGenerator
 edu.berkeley.cs.jqf.examples.imageio.ImageInputStreamGenerator
 edu.berkeley.cs.jqf.examples.jgrapht.ModelBasedDirectedGraphGenerator
 edu.berkeley.cs.jqf.examples.jgrapht.ModelBasedDirectedGraphGenerator
