+++ fuzz/src/main/java/edu/berkeley/cs/jqf/fuzz/junit/quickcheck/InputStreamGenerator.java
@@ -26,7 +26,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package edu.berkeley.cs.jqf.examples.common;
+package edu.berkeley.cs.jqf.fuzz.junit.quickcheck;
 
 import java.io.EOFException;
 import java.io.IOException;
@@ -37,6 +37,16 @@ import com.pholser.junit.quickcheck.generator.Generator;
 import com.pholser.junit.quickcheck.random.SourceOfRandomness;
 
 /**
+ * Lazy provider of bytes from an input stream. This is useful
+ * when you want the parameter sequence to be exactly equal to
+ * the raw input, such as when fuzzing with AFL.
+ *
+ * Use this only with AFL-like front-ends, which return
+ * bytes from a fixed-size file. This generator may not work
+ * properly with front-ends like Zest that generate fresh
+ * values at the end of a parameter sequence, since that would
+ * make the InputStream infinitely long.
+ *
  * @author Rohan Padhye
  */
 public class InputStreamGenerator extends Generator<InputStream> {
@@ -49,6 +59,10 @@ public class InputStreamGenerator extends Generator<InputStream> {
         return new InputStream() {
            @Override
            public int read() throws IOException {
+               // Keep asking for new random bytes until the
+               // SourceOfRandomness runs out of parameters. This is designed
+               // to work with fixed-size parameter sequences, such as when
+               // fuzzing with AFL.
                try {
                    byte nextByte = sourceOfRandomness.nextByte(Byte.MIN_VALUE, Byte.MAX_VALUE);
                    int nextInt = nextByte & 0xFF;
