+++ fuzz/src/main/resources/META-INF/services/com.pholser.junit.quickcheck.generator.Generator
@@ -1,0 +1,2 @@
+edu.berkeley.cs.jqf.fuzz.junit.quickcheck.InputStreamGenerator
+
